requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
Django==3.1
djangorestframework==3.11.1
gunicorn==20.0.4
lxml==4.5.2
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
